+++ src/components/Search/useDebouncedFunction.tsx
@@ -1,0 +1,23 @@
+import { useEffect, useRef } from 'react';
+
+export default function useDebouncedFunction(
+  func: any,
+  delay: number,
+  cleanUp = false
+) {
+  const timeoutRef: any = useRef();
+
+  function clearTimer() {
+    if (timeoutRef.current) {
+      clearTimeout(timeoutRef.current);
+      timeoutRef.current = undefined;
+    }
+  }
+
+  useEffect(() => (cleanUp ? clearTimer : undefined), [cleanUp]);
+
+  return (...args: any) => {
+    clearTimer();
+    timeoutRef.current = setTimeout(() => func(...args), delay);
+  };
+}
